+++ test/domain/initial_setting/initial_setting_store_test.dart
@@ -9,7 +9,7 @@ import 'package:pilll/entity/setting.codegen.dart';
 import 'package:pilll/service/auth.dart';
 import 'package:pilll/service/day.dart';
 import 'package:pilll/service/pill_sheet.dart';
-import 'package:pilll/service/pill_sheet_group.dart';
+import 'package:pilll/database/pill_sheet_group.dart';
 import 'package:pilll/service/pill_sheet_modified_history.dart';
 import 'package:pilll/service/setting.dart';
 import 'package:pilll/service/user.dart';
